requirements.txt
amqp==5.3.1
attrs==25.3.0
Automat==25.4.16
billiard==4.2.1
celery==5.5.3
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
constantly==23.10.4
cryptography==45.0.4
cssselect==1.3.0
defusedxml==0.7.1
filelock==3.18.0
flower==2.0.1
greenlet==3.2.3
humanize==4.12.3
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
itemadapter==0.11.0
itemloaders==1.3.2
jmespath==1.0.1
kombu==5.5.4
lxml==5.4.0
packaging==25.0
parsel==1.10.0
prometheus_client==0.22.1
prompt_toolkit==3.0.51
Protego==0.4.0
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
PyDispatcher==2.0.7
pyOpenSSL==25.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
queuelib==1.8.0
redis==6.2.0
requests==2.32.4
requests-file==2.1.0
Scrapy==2.13.2
service-identity==24.2.0
setuptools==80.9.0
six==1.17.0
SQLAlchemy==2.0.41
tldextract==5.3.0
tornado==6.5.1
Twisted==25.5.0
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.5.0
vine==5.1.0
w3lib==2.3.1
wcwidth==0.2.13
zope.interface==7.2
